+++ bento/__init__.py
@@ -8,3 +8,28 @@ from . import datasets as ds
 from . import io
 from .plotting import _colors as colors
 from ._constants import CosMx, Merscope, Xenium
+
+from .points import (
+    distance_stats,
+    polarity,
+    moments,
+    density as points_density,
+    morans_i,
+    ripley,
+)
+from .shapes import (
+    area,
+    aspect_ratio,
+    bounds,
+    centroid,
+    opening,
+    perimeter,
+    radius,
+    second_moment,
+    span,
+)
+from .images import (
+    total_intensity,
+    mean_intensity,
+    regionprops,
+)
