--- bento/experimental/__init__.py
@@ -1,3 +1,0 @@
-from . import points as pts
-from . import shapes as shp
-from . import images as img
