--- bento/experimental/images/__init__.py
@@ -1,1 +1,0 @@
-from ._measure import measure, total_intensity, mean_intensity, regionprops
