--- bento/experimental/points/__init__.py
@@ -1,9 +1,0 @@
-from ._api import (
-    distance_stats,
-    polarity,
-    density,
-    moments,
-    morans_i,
-    ripley,
-)
-from ._cluster import cluster
