--- bento/experimental/shapes/__init__.py
@@ -1,1 +1,0 @@
-from ._measure import aspect_ratio, radius, span, second_moment, opening, area, perimeter, bounds, centroid
