+++ bento/images/__init__.py
@@ -1,0 +1,11 @@
+from .total_intensity import total_intensity
+from .mean_intensity import mean_intensity
+from .regionprops import regionprops
+
+__all__ = [
+    "total_intensity",
+    "mean_intensity",
+    "regionprops",
+]
+
+
