+++ bento/shapes/__init__.py
@@ -1,0 +1,23 @@
+from .area import area
+from .aspect_ratio import aspect_ratio
+from .bounds import bounds
+from .centroid import centroid
+from .opening import opening
+from .perimeter import perimeter
+from .radius import radius
+from .second_moment import second_moment
+from .span import span
+
+__all__ = [
+    "area",
+    "aspect_ratio",
+    "bounds",
+    "centroid",
+    "opening",
+    "perimeter",
+    "radius",
+    "second_moment",
+    "span",
+]
+
+
