+++ bento/tools/__init__.py
@@ -2,38 +2,22 @@ from ._colocation import coloc_quotient, colocation
 from ._composition import comp, comp_diff
 from ._flux import flux, fluxmap
 from ._flux_enrichment import fe, fe_fazal2019, fe_xia2019, gene_sets, load_gene_sets
-from ._lp import lp, lp_stats, lp_diff_discrete, lp_diff_continuous
-from ._point_features import (
-    analyze_points, 
-    list_point_features, 
-    register_point_feature,
-    PointFeature,
-    ShapeProximity,
-    ShapeAsymmetry,
-    PointDispersionNorm,
-    ShapeDispersionNorm,
-    ShapeDistance,
-    ShapeOffset,
-    PointDispersion,
-    ShapeDispersion,
-    RipleyStats,
-    ShapeEnrichment,
-
-)
-from ._shape_features import (
-    analyze_shapes,
-    shape_stats,
-    register_shape_feature,
-    list_shape_features,
-    area,
-    aspect_ratio,
-    bounds,
-    density,
-    opening,
-    perimeter,
-    radius,
-    raster,
-    second_moment,
-    span,
-)
 from ._decomposition import decompose
+
+# Explicit export surface
+__all__ = [
+    # Composition/colocation/flux
+    "coloc_quotient",
+    "colocation",
+    "comp",
+    "comp_diff",
+    "flux",
+    "fluxmap",
+    "fe",
+    "fe_fazal2019",
+    "fe_xia2019",
+    "gene_sets",
+    "load_gene_sets",
+    # Other
+    "decompose",
+]
